% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDefs.R
\name{add.defs}
\alias{add.defs}
\title{Add single row to definitions table}
\usage{
add.defs(dtDefs, id = "", varname = "", formula = "", variance = 0,
  min = 0, max = 0, dist = "Normal", link = "identity", cMethod = "",
  nIperC = 0, nClust = 0, nTrt = 0, varcat = "", missType = "")
}
\arguments{
\item{dtDefs}{Definition data.table to be modified}

\item{id}{Name of field that contains idnetifier}

\item{formula}{test}

\item{variance}{test}

\item{min}{test}

\item{max}{test}

\item{dist}{test}

\item{link}{test}

\item{cMethod}{test}

\item{nIperC}{test}

\item{nClust}{test}

\item{nTrt}{test}

\item{varcat}{test}

\item{missType}{test}

\item{varName}{test}
}
\value{
A data.table named dtName that is an updated data defnitions table
}
\description{
Add single row to definitions table
}

