% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDefs.R
\name{addDefs}
\alias{addDefs}
\title{Add single row to definitions table}
\usage{
addDefs(dtDefs, varname = "", formula = "", variance = 0,
  dist = "Normal", link = "identity", nTrt = 0, cMethod = "",
  missType = "")
}
\arguments{
\item{dtDefs}{Definition data.table to be modified}

\item{formula}{test}

\item{variance}{test}

\item{dist}{test}

\item{link}{test}

\item{nTrt}{test}

\item{cMethod}{test}

\item{missType}{test}

\item{varName}{test}
}
\value{
A data.table named dtName that is an updated data defnitions table
}
\description{
Add single row to definitions table
}

