% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDt.R
\name{addDt}
\alias{addDt}
\title{Add columns to existing data set}
\usage{
addDt(dtDefs, dtOld)
}
\arguments{
\item{dtDefs}{Name of definitions for added columns}

\item{dtOld}{Name of data table that is to be updated}
}
\value{
An updated data.table that contains the added simulated data
}
\description{
Add columns to existing data set
}

