% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPeriods.R
\name{addPeriods}
\alias{addPeriods}
\title{Create longitudinal/panel data}
\usage{
addPeriods(dtName, nPeriods = NULL, idvars = "id", timevars = NULL,
  timevarName = "timevar", timeid = "timeID")
}
\arguments{
\item{dtName}{Name of existing data table}

\item{nPeriods}{Number of time periods for each record}

\item{idvars}{Names of index variables (in a string vector) that will be
repeated during each time period}

\item{timevars}{Names of time dependent variables. Defaults to NULL.}

\item{timevarName}{Name of new time dependent variable}

\item{timeid}{Variable name for new index field. Defaults to "timevar"}
}
\value{
An updated data.table that that has multiple rows
per observation in dtName
}
\description{
Create longitudinal/panel data
}

