% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_addStrataCode.R
\name{addStrataCode}
\alias{addStrataCode}
\title{Assign treatment}
\usage{
addStrataCode(dt, strata)
}
\arguments{
\item{dt}{data table}

\item{strata}{vector of string names representing strata}
}
\value{
An integer (group) ranging from 1 to length of the probability vector
}
\description{
Assign treatment
}

