% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_data.R
\name{addSynthetic}
\alias{addSynthetic}
\title{Add synthetic data}
\usage{
addSynthetic(
  dtOld,
  dtFrom,
  vars = names(dtFrom)[names(dtFrom) != id],
  id = "id"
)
}
\arguments{
\item{dtOld}{data.table that is to be modified}

\item{dtFrom}{Data table that contains the source data}

\item{vars}{A vector of string names specifying the fields that will be
sampled. The default is that all variables will be selected.}

\item{id}{A string specifying the field that serves as the record id. The
default field is "id".}
}
\value{
A data.table that contains the added synthetic data.
}
\description{
Add synthetic data
}
\examples{
### Create fake "real" data set

d <- defData(varname = "a", formula = 3, variance = 1, dist = "normal")
d <- defData(d, varname = "b", formula = 5, dist = "poisson")
d <- defData(d, varname = "c", formula = 0.3, dist = "binary")
d <- defData(d, varname = "d", formula = "a + b + 3*c", variance = 2, dist = "normal")

A <- genData(1000, d)

### Create synthetic data set from "observed" data set A:

def <- defData(varname = "x", formula = 0, variance = 5)

S <- genData(120, def)
S <- addSynthetic(dtOld = S, dtFrom = A, vars = c("b", "d"))
}
\concept{generate_data}
