% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{arm64test}
\alias{arm64test}
\title{arm64 test}
\usage{
arm64test(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
A third number.
}
\description{
arm64 test
}
\examples{
arm64test(7, 1)
}
\concept{utility}
