% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defDataAdd.R
\name{defDataAdd}
\alias{defDataAdd}
\title{Add single row to definitions table that will be used to add data to an
existing data.table}
\usage{
defDataAdd(dtDefs = NULL, varname, formula, variance = 0, dist = "Normal",
  link = "identity")
}
\arguments{
\item{dtDefs}{Definition data.table to be modified}

\item{formula}{An R expression for mean (string)}

\item{variance}{Number}

\item{dist}{Distribution. For possibilities, see details}

\item{link}{The link function for the mean, see details}

\item{varName}{Name (string) of new variable}
}
\value{
A data.table named dtName that is an updated data defnitions table
}
\description{
Add single row to definitions table that will be used to add data to an
existing data.table
}

