% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defMiss.R
\name{defMiss}
\alias{defMiss}
\title{Add single row to definitions table for missing data}
\usage{
defMiss(dtDefs = NULL, varname, formula, logit.link = FALSE,
  baseline = FALSE, monotonic = FALSE)
}
\arguments{
\item{dtDefs}{Definition data.table to be modified}

\item{varname}{Name of variable with missingness}

\item{formula}{Formula to describe pattern of missingness}

\item{logit.link}{Indicator set to TRUE when the probability of missingness
is based on a logit model.}

\item{baseline}{Indicator is set to TRUE if the variable is a baseline
measure and should be missing throughout an entire observation period. This
is applicable to repeated measures/longitudinal data.}

\item{monotonic}{Indicator set to TRUE if missingness at time t is followed
by missingness at all follow-up times > t.}
}
\value{
A data.table named dtName that is an updated data defnitions table
}
\description{
Add single row to definitions table for missing data
}

