% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defRead.R
\name{defRead}
\alias{defRead}
\title{Read external csv data set definitions}
\usage{
defRead(filen, id = "id")
}
\arguments{
\item{filen}{String file name, including full path. Must be a csv file.}

\item{id}{string that includes name of id field. Defaults to "id"}
}
\value{
A data.table with data set definitions
}
\description{
Read external csv data set definitions
}

