% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defReadAdd.R
\name{defReadAdd}
\alias{defReadAdd}
\title{Read external csv data set definitions for adding columns}
\usage{
defReadAdd(filen)
}
\arguments{
\item{filen}{String file name, including full path. Must be a csv file.}
}
\value{
A data.table with data set definitions
}
\description{
Read external csv data set definitions for adding columns
}

