% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defSurv.R
\name{defSurv}
\alias{defSurv}
\title{Add single row to survival definitions}
\usage{
defSurv(dtDefs = NULL, varname, formula = 0, scale, shape = 1)
}
\arguments{
\item{dtDefs}{Definition data.table to be modified}

\item{formula}{Covariates predicting survival}

\item{scale}{Scale parameter for the Weibull distribution.}

\item{shape}{The shape of the Weibull distribution. Shape = 1 for
an exponential distriubtion}

\item{varName}{Variable name}
}
\value{
A data.table named dtName that is an updated data defnitions table
}
\description{
Add single row to survival definitions
}

