% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{delColumns}
\alias{delColumns}
\title{Delete columns from existing data set}
\usage{
delColumns(dtOld, vars)
}
\arguments{
\item{dtOld}{Name of data table that is to be updated.}

\item{vars}{Vector of column names (as strings).}
}
\value{
An updated data.table without \code{vars}.
}
\description{
Delete columns from existing data set
}
\examples{
# New data set

def <- defData(varname = "x", dist = "noZeroPoisson", formula = 7, id = "idnum")
def <- defData(def, varname = "xUni", dist = "uniformInt", formula = "x-3;x+3")

dt <- genData(10, def)
dt

# Delete column

dt <- delColumns(dt, "x")
dt
}
\concept{utility}
