% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_genPseudo.R
\name{.genPseudoSeq}
\alias{.genPseudoSeq}
\title{Pseudo-Random Sequence}
\usage{
.genPseudoSeq(n, formula, link)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{formula}{String that specifies the pool of numbers to be sampled from
and the number of repetions. "pool;reps" e.g. "1,2,4,5;4"}

\item{link}{String \code{"fill"} to indicate if a length mismatch should be
tolerated, anything else will enforce \code{n = reps * length(vector)}.}
}
\value{
A data.frame column (aka vector) with the simulated data
}
\description{
Generate a sequence of \code{n/length(vector)} randomly ordered repetitions
of a numeric vector.
}
\keyword{internal}
