% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_strings.R
\name{ensureReference}
\alias{ensureReference}
\title{Function to check if the option 'simstudy.use_reference' is set
which signals that data.tables should not be copied in generation
 processes. (see issue #50)}
\usage{
ensureReference(dt)
}
\description{
Function to check if the option 'simstudy.use_reference' is set
which signals that data.tables should not be copied in generation
 processes. (see issue #50)
}
