% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_genbinom.R
\name{genbinom}
\alias{genbinom}
\title{Internal function called by generate - returns Binary data}
\usage{
genbinom(n, formula, link, dtSim)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{link}{Link function: Identity or Logit}

\item{dtSim}{Incomplete simulated data.table}

\item{forumula}{String that specifies the formula for the probability}
}
\value{
A data.frame column  with the updated simulated data
}
\description{
Internal function called by generate - returns Binary data
}

