% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_correlated_data.R
\name{genBlockMat}
\alias{genBlockMat}
\title{Create a block correlation matrix}
\usage{
genBlockMat(rho, nInds, nPeriods, corstr = "ind", iRho = NULL)
}
\arguments{
\item{rho}{The correlation coefficient between 0 and 1.}

\item{nInds}{The number of units (individuals) in each period.}

\item{nPeriods}{The number periods that data are observed.}

\item{corstr}{Correlation structure for the matrix. Options include "ind" 
for an independence structure (default), cs" for a compound symmetry 
structure and "ar1" for an autoregressive structure.}

\item{iRho}{Optional second correlation coefficient for within-individual across-period
correlation in case the measurements for individuals are repeated over time. 
This will use the same correlation structure as the between-individual correlation
specified by "corstr".}
}
\value{
A block correlation matrix of size (nInds \* nPeriods) x (nInds \* nPeriods)
}
\description{
The function genBlockMat() generates correlation matrices that 
can accommodate clustered observations over time where the within-cluster 
between-individual correlation in the same time period can be different from the 
within-cluster between-individual correlation across time periods.The matrix
generated here can be used in function addCorGen().
}
\examples{
genBlockMat(rho = 0.6, nInds = 2, nPeriods = 3, corstr = "ind")
genBlockMat(rho = 0.6, nInds = 2, nPeriods = 3, corstr = "cs")
genBlockMat(rho = 0.6, nInds = 3, nPeriods = 3, corstr = "ar1")

genBlockMat(rho = 0.6, nInds = 3, nPeriods = 3, corstr = "cs", iRho = 0.8)
genBlockMat(rho = 0.6, nInds = 3, nPeriods = 3, corstr = "ar1", iRho = 0.8)
}
\concept{correlated}
