% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_gencat.R
\name{gencat}
\alias{gencat}
\title{Internal function called by generate - returns categorical data}
\usage{
gencat(n, formula)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{formula}{String that specifies the probabilities}
}
\value{
A data.frame column with the updated simulated data
}
\description{
Internal function called by generate - returns categorical data
}

