% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_genclust.R
\name{genclust}
\alias{genclust}
\title{Internal function called by generate - returns cluster level}
\usage{
genclust(cMethod, nIperC, nClust)
}
\arguments{
\item{cMethod}{Method of allocating subjects to treatment group. Can be
equal assingment or variable assignment based on covariate or 
probability. Variable assignment currently works for binary only.}

\item{nIperC}{Number subjects per cluster}

\item{nClust}{Number of clusters}
}
\value{
A data.frame column with cluster assignments
}
\description{
Internal function called by generate - returns cluster level
}

