% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCorData.R
\name{genCorData}
\alias{genCorData}
\title{Create correlated data}
\usage{
genCorData(n, nvars, mu, sigma, rho, corstr = "ind", cnames = NULL,
  idname = "id")
}
\arguments{
\item{n}{Number of observations}

\item{nvars}{Number of columns of correlated data}

\item{mu}{A vector of means. The length of mu must be nvars.}

\item{sigma}{Common variance shared by all variables}

\item{rho}{Correlation coefficient, -1 <= rho <= 1.}

\item{corstr}{Correlation structure of the variance-covariance matrix
defined by sigma and rho. Options include "ind" for an independence 
structure, "cs" for a compound symmetry structure, and "ar1" for an
autoregressive structure.}

\item{cnames}{Explicit column names. A single string with names separated
by commas. If no string is provided, the default names will be V#, where #
represents the column.}

\item{idname}{The name of the index id name. Defaults to "id."}
}
\value{
A data.table with n rows and nvars + 1 columns
}
\description{
Create correlated data
}

