% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_correlated_data.R
\name{genCorMat}
\alias{genCorMat}
\title{Create a correlation matrix}
\usage{
genCorMat(nvars, cors = NULL, rho = NULL, corstr = "cs", nclusters = 1)
}
\arguments{
\item{nvars}{number of rows and columns (i.e. number of variables) for correlation matrix}

\item{cors}{vector of correlations.}

\item{rho}{Correlation coefficient, -1 <= rho <= 1. Use if corMatrix is not provided.}

\item{corstr}{Correlation structure of the variance-covariance matrix
defined by sigma and rho. Options include "cs" for a compound symmetry structure
and "ar1" for an autoregressive structure.}

\item{nclusters}{An integer that indicates the number of matrices that will be generated.}
}
\value{
correlation matrix of size nvars x nvars
}
\description{
Create a correlation matrix
}
\details{
If the vector cors is not specified, a random correlation matrix is generated with no assumptions.
If the vector is provided, it should be interpreted as the lower triangle of the correlation
matrix, and is specified by reading down the columns. For example, if CM is the correlation matrix and
nvars = 3, then CM[2,1] = cors[1],  CM[3,1] = cors[2], and CM[3,2] = cors[3].
}
\examples{
genCorMat(3, c(.3, -.2, .1))
genCorMat(3)

genCorMat(5, c(.3, -.2, .1, .2, .5, .2, -.1, .3, .1, .2))
genCorMat(5)
}
\concept{correlated}
