% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_gendeterm.R
\name{gendeterm}
\alias{gendeterm}
\title{Internal function called by generate - returns non-random data}
\usage{
gendeterm(n, formula, dtSim)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{formula}{String that specifies the formula for the mean}

\item{dtSim}{Incomplete simulated data.table}
}
\value{
A data.frame column  with the updated simulated data
}
\description{
Internal function called by generate - returns non-random data
}

