% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendt.R
\name{gendt}
\alias{gendt}
\title{Calling function to simulate data}
\usage{
gendt(dtDefs, n)
}
\arguments{
\item{dtDefs}{name of definitions data.table/data.frame}

\item{n}{the number (integer) of observations required in the data set}
}
\value{
A data.table that contains the simulated data
}
\description{
Calling function to simulate data
}

