% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gendtClust.R
\name{gendtClust}
\alias{gendtClust}
\title{Simulate data set that is one layer down in a multilevel data context}
\usage{
gendtClust(dtDefs, dtClust, cLevelVar, numIndsVar)
}
\arguments{
\item{dtDefs}{dtDefs Name of definitions data.table/data.frame}

\item{dtClust}{Name of cluster level data}

\item{cLevelVar}{Variable name (string) of cluster id in dtClust}

\item{numIndsVar}{Variable name (string) of number of observations 
per cluster in dtClust}
}
\value{
A simulated data table
}
\description{
Simulate data set that is one layer down in a multilevel data context
}

