% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_strings.R
\name{generate}
\alias{generate}
\title{Generate Data}
\usage{
generate(def, env = parent.frame())
}
\arguments{
\item{def}{Row of a definition table containing expressions with the elements of the definition:
`dtDefs`, `varname`, `formula`, `param1`, `param2`, `data`}
}
\description{
Each distribution has a generate method, that generates data.
The expected format is...
}
\concept{custom_distributions}
