% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_genMissDataMat.R
\name{genMissDataMat}
\alias{genMissDataMat}
\title{Internal function called by genMiss - returns a missing data matrix}
\usage{
genMissDataMat(dtName, dtTemp, idvars, missDefs)
}
\arguments{
\item{dtName}{Name of complete data set}

\item{dtTemp}{Name of data set with unique ids only}

\item{idvars}{To be filled in}

\item{missDefs}{To be filled in}
}
\value{
A missing data matrix of 0/1, where 1 indicates missing
}
\description{
Internal function called by genMiss - returns a missing data matrix
}

