% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_gennorm.R
\name{gennorm}
\alias{gennorm}
\title{Internal function called by generate - returns Normal data}
\usage{
gennorm(n, formula, variance, link, dtSim)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{variance}{Variance of the normal distribution}

\item{link}{Link function not used in normal distribution}

\item{dtSim}{Incomplete simulated data.table}

\item{forumula}{String that specifies the formula for the mean}
}
\value{
A data.frame column  with the updated simulated data
}
\description{
Internal function called by generate - returns Normal data
}

