% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genObs.R
\name{genObs}
\alias{genObs}
\title{Create an observed data set that includes missing data}
\usage{
genObs(dtName, dtMiss, idvars)
}
\arguments{
\item{dtName}{Name of complete data set}

\item{dtMiss}{Name of missing data matrix}

\item{idvars}{Index variables that cannot be missing}
}
\value{
A data table that represents observed data, including 
missing data
}
\description{
Create an observed data set that includes missing data
}

