% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{genOrdCatNP}
\alias{genOrdCatNP}
\title{Generate ordinal categorical data without proportionality assumption}
\usage{
genOrdCatNP(
  dtName,
  adjVar = NULL,
  baseprobs,
  npVar = NULL,
  npAdj = NULL,
  catVar = "cat",
  asFactor = TRUE,
  idname = "id",
  prefix = "grp",
  rho = 0,
  corstr = "ind",
  corMatrix = NULL
)
}
\arguments{
\item{dtName}{Name of complete data set}

\item{adjVar}{Adjustment variable  name in dtName - determines
logistic shift. This is specified assuming a cumulative logit
link.}

\item{baseprobs}{Baseline probability expressed as a vector or matrix of
probabilities. The values (per row) must sum to <= 1. If \code{rowSums(baseprobs) < 1}, an additional category is added with probability \code{1 - rowSums(baseprobs)}. The number of rows represents the number of new
categorical variables. The number of columns represents the number of
possible responses - if an particular category has fewer possible responses,
assign zero probability to non-relevant columns.}

\item{npVar}{Name(s) of field(s) that break the non-proportionality assumption.
A string or vector of strings.}

\item{npAdj}{Vector or matrix of adjustments. Length of vector or number of
columns must be the same number of response levels. The last column will be
zero regardless of what is entered.}

\item{catVar}{Name of the new categorical field. Defaults to "cat". Can be a
character vector with a name for each new variable defined via \code{baseprobs}.
Will be overridden by \code{prefix} if more than one variable is defined and
\code{length(catVar) == 1}.}

\item{asFactor}{If \code{asFactor == TRUE} (default), new field is returned
as a factor. If \code{asFactor == FALSE}, new field is returned as an integer.}

\item{idname}{Name of the id column in \code{dtName}.}

\item{prefix}{A string. The names of the new variables will be a
concatenation of the prefix and a sequence of integers indicating the
variable number.}

\item{rho}{Correlation coefficient, -1 < rho < 1. Use if corMatrix is not
provided.}

\item{corstr}{Correlation structure of the variance-covariance matrix defined
by sigma and rho. Options include "ind" for an independence structure, "cs"
for a compound symmetry structure, and "ar1" for an autoregressive structure.}

\item{corMatrix}{Correlation matrix can be entered directly. It must be
symmetrical and positive definite. It is not a required field; if a matrix is
not provided, then a structure and correlation coefficient rho must be
specified. (The matrix created via \code{rho} and \code{corstr} must also be positive
definite.)}
}
\value{
Original data.table with added categorical field.
}
\description{
Ordinal categorical data without the proportionality assumption is
added to an existing data set.
}
\examples{
# Ordinal Categorical Data without proportionality assumption ----

d1 <- defData(varname = "z", formula = 0, variance = .25)

dd <- genData(10, d1)
dd <- trtAssign(dd, grpName = "rx")

baseprobs <- c(.4, .3, .2, .1)
npAdj <- c(0, 0.1, 0.3, 0)

dd <- genOrdCatNP(dtName = dd, adjVar = "z", baseprobs = baseprobs, 
                 npVar = "rx", npAdj = npAdj)

}
\concept{categorical}
\concept{correlated}
\concept{generate_data}
