% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_genpois.R
\name{genpois}
\alias{genpois}
\title{Internal function called by generate - returns Poisson count data}
\usage{
genpois(n, formula, link, dtSim)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{formula}{String that specifies the formula for the mean}

\item{link}{Link function: Identity or Log}

\item{dtSim}{Incomplete simulated data.table}
}
\value{
A data.frame column with the updated simulated data
}
\description{
Internal function called by generate - returns Poisson count data
}

