% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_genRx.R
\name{genRx}
\alias{genRx}
\title{Internal function called by generate - returns treatment assignment}
\usage{
genRx(n, formula, link, cMethod, nTrt = 2, dtSim)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{formula}{String that specifies the formula for the mean}

\item{link}{Link function: Identity or Logit}

\item{cMethod}{Method of allocating subjects to treatment group. Can be
equal assingment or variable assignment based on covariate or 
probability. Variable assignment currently works for binary only.}

\item{nTrt}{Number of treatment groups}

\item{dtSim}{Incomplete simulated data.table}
}
\value{
A data.frame with the updated simulated data
}
\description{
Internal function called by generate - returns treatment assignment
}

