% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSurv.R
\name{genSurv}
\alias{genSurv}
\title{Generate missing data}
\usage{
genSurv(dtName, survDefs)
}
\arguments{
\item{dtName}{Name of complete data set}

\item{survDefs}{Definitions of survival}
}
\value{
Original matrix with survival time
}
\description{
Generate missing data
}

