% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{genSynthetic}
\alias{genSynthetic}
\title{Generate synthetic data}
\usage{
genSynthetic(dtFrom, n = nrow(dtFrom), vars = NULL, id = "id")
}
\arguments{
\item{dtFrom}{Data set that contains the source data}

\item{n}{Number of samples to draw from the source data. The default
is number of records that are in the source data file.}

\item{vars}{A vector of string names specifying the fields that will be
sampled. The default is that all variables will be selected.}

\item{id}{A string specifying the field that serves as the record id. The
default field is "id".}
}
\value{
A data table with the generated data
}
\description{
Synthetic data is generated from an existing data set
}
\examples{
# Baseline data definitions

}
\concept{generate_data}
