% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_genunif.R
\name{genunif}
\alias{genunif}
\title{Internal function called by generate - returns Uniform data}
\usage{
genunif(n, formula)
}
\arguments{
\item{n}{The number of observations required in the data set}

\item{min}{Lower value of uniform distribution}

\item{max}{Upper value of uniform distribution}
}
\value{
A data.frame column  with the updated simulated data
}
\description{
Internal function called by generate - returns Uniform data
}

