% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getBeta0}
\alias{getBeta0}
\title{Determine intercept parameter for logistic regression with target prevalence}
\usage{
getBeta0(targetProp, def, coefs, intLow = -10, intHigh = 10, tolerance = 1e-05)
}
\arguments{
\item{targetProp}{The target proportion. A value between 0 and 1.}

\item{def}{A data definition table for the covariates in the underlying
population.}

\item{coefs}{A vector of coefficients that reflect the relationship between 
each of the covariates and the log-odds of the outcome.}

\item{intLow}{The low endpoint for the search range. Defaults to -10.}

\item{intHigh}{The high endpoint for the search range. Defaults to 10.}

\item{tolerance}{The minimum stopping distance between the adjusted low and high
endpoints. Defaults to 0.00001.}
}
\value{
A value representing th desired intercept parameter for the logistic
model.
}
\description{
An iterative bisection procedure that can be used to determine 
the numeric value of the intercept parameter for a data generating process 
(based on a logistic regression model) that has a specific target population 
prevalence of a binary outcome.
}
\examples{
d1 <- defData(varname = "x1", formula = 0, variance = 1)
d1 <- defData(d1, varname = "b1", formula = 0.5, dist = "binary")

coefs <- log(c(1.2, 0.8))

getBeta0(targetProp = 0.11, def = d1, coefs = coefs)

}
\references{
Austin, Peter C. "The iterative bisection procedure: a useful 
tool for determining parameter values in data-generating processes in 
Monte Carlo simulations." BMC Medical Research Methodology 23, 
no. 1 (2023): 1-10.
}
\concept{utility}
