% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getBetas}
\alias{getBetas}
\title{Determine parameters for logistic regression with target prevalence and 
risk ratio or risk difference}
\usage{
getBetas(
  targetPrev,
  targetComp,
  def,
  coefs,
  comparison = "rr",
  intLow = -10,
  intHigh = 10,
  tolerance = 1e-04,
  sampleSize = 5e+05,
  trtName = "A"
)
}
\arguments{
\item{targetPrev}{The target population prevalence. A value between 0 and 1.}

\item{targetComp}{The target risk ratio or risk difference. The risk ratio can be
any positive value, and the risk difference is in any value between -1 and 1.}

\item{def}{A data definition table for the covariates in the underlying
population.}

\item{coefs}{A vector of coefficients that reflect the relationship between 
each of the covariates and the log-odds of the outcome.}

\item{comparison}{This indicates whether the target is a risk ratio ("rr") 
or risk difference ("rd"). The default value is "rr."}

\item{intLow}{The low endpoint for the search range. Defaults to -10.}

\item{intHigh}{The high endpoint for the search range. Defaults to 10.}

\item{tolerance}{The minimum stopping distance between the adjusted low and high
endpoints. Defaults to 0.0001.}

\item{sampleSize}{The number of units to generate for the bisection algorithm. 
The default is 5e+05. To get a reliable estimate, the value 
should be no smaller than 1e+05.}

\item{trtName}{The name of the variable treatment/exposure variable. Defaults 
to "A".}
}
\value{
A vector of parameters including the intercept, treatment/exposure
effect, and covariate coefficients for the logistic model data generating 
process.
}
\description{
An iterative bisection procedure that can be used to determine 
the numeric value of the intercept and treatment effect parameter for a 
data generating process (based on a logistic regression model) that has a 
specific target population prevalence of a binary outcome, and a target risk
ratio or risk difference.
}
\examples{
d1 <- defData(varname = "x1", formula = 0, variance = 1)
d1 <- defData(d1, varname = "b1", formula = 0.5, dist = "binary")

coefs <- log(c(1.2, 0.8))

getBetas(targetPrev = 0.10, targetComp = 1.2, def = d1, coefs = coefs)

}
\references{
Austin, Peter C. "The iterative bisection procedure: a useful 
tool for determining parameter values in data-generating processes in 
Monte Carlo simulations." BMC Medical Research Methodology 23, 
no. 1 (2023): 1-10.
}
\concept{utility}
