% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initDefs.R
\name{initDefs}
\alias{initDefs}
\title{Create initial definitions table}
\usage{
initDefs(id = "id")
}
\arguments{
\item{id}{A string with the variable that contains the id identifier}
}
\value{
A data.table that has zeros and blanks
}
\description{
Create initial definitions table
}

