% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logisticCoefs}
\alias{logisticCoefs}
\title{Determine intercept parameter for logistic regression with target prevalence}
\usage{
logisticCoefs(
  defCov,
  coefs,
  popPrev,
  rr = NULL,
  rd = NULL,
  auc = NULL,
  tolerance = 0.001,
  sampleSize = 5e+05,
  trtName = "A"
)
}
\arguments{
\item{defCov}{A data definition table for the covariates in the underlying
population. This specifies the underlying distribution of the covariates.}

\item{coefs}{A vector of coefficients that reflect the relationship between 
each of the covariates and the log-odds of the outcome.}

\item{popPrev}{The target population prevalence of the outcome. 
A value between 0 and 1.}

\item{rr}{If the target statistic is a risk ratio, a value between 0 and
1/popPrev must be provided. Defaults to NULL.}

\item{rd}{If the target statistic is a risk difference, a value between
-(popPrev) and (1 - popPrev) must be provided. Defaults to NULL}

\item{auc}{If the target statistic is the AUC, a value between 0.5 and 1.0 
must be provided. Defaults to NULL.}

\item{tolerance}{The minimum stopping distance between the adjusted low and high
endpoints. Defaults to 0.001.}

\item{sampleSize}{The number of units to generate for the bisection algorithm. 
The default is 5e+05. To get a reliable estimate, the value 
should be no smaller than 1e+05.}

\item{trtName}{If either a risk ratio or risk difference is the target statistic,
a treatment/exposure variable name can be provided. trtName defaults to "A".}
}
\value{
A vector of parameters including the intercept and covariate 
coefficients for the logistic model data generating process.
}
\description{
An iterative bisection procedure that can be used to determine 
the numeric value of the intercept parameter for a data generating process 
(based on a logistic regression model) that has a specific target population 
prevalence of a binary outcome.
}
\examples{
d1 <- defData(varname = "x1", formula = 0, variance = 1)
d1 <- defData(d1, varname = "b1", formula = 0.5, dist = "binary")

coefs <- log(c(1.2, 0.8))

logisticCoefs(d1, coefs, popPrev = 0.20) 
logisticCoefs(d1, coefs, popPrev = 0.20, rr = 1.50, trtName = "rx") 
logisticCoefs(d1, coefs, popPrev = 0.20, rd = 0.30, trtName = "rx")
logisticCoefs(d1, coefs, popPrev = 0.20, auc = 0.80)
}
\references{
Austin, Peter C. "The iterative bisection procedure: a useful 
tool for determining parameter values in data-generating processes in 
Monte Carlo simulations." BMC Medical Research Methodology 23, 
no. 1 (2023): 1-10.
}
\concept{utility}
