% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_loProb.R
\name{loProb}
\alias{loProb}
\title{Internal function convert log odds to probability}
\usage{
loProb(logodds)
}
\arguments{
\item{logodds}{Log odds}
}
\value{
Probability
}
\description{
Internal function convert log odds to probability
}

