% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge two data tables}
\usage{
mergeData(dt1, dt2, idvars)
}
\arguments{
\item{dt1}{Name of first data.table}

\item{dt2}{Name of second data.table}

\item{idvars}{Vector of string names to merge on}
}
\value{
A new data table that merges dt2 with dt1
}
\description{
Merge two data tables
}

