% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeLevels.R
\name{mergeLevels}
\alias{mergeLevels}
\title{Merge two data tables - used to merge multilevel data}
\usage{
mergeLevels(dt2, dt1, idvars)
}
\arguments{
\item{dt2}{Name of level 2 data set}

\item{dt1}{Name of level 1 data set}

\item{idvars}{Vector of string names to merge on}
}
\value{
A new data table that merges dt2 with dt1
}
\description{
Merge two data tables - used to merge multilevel data
}

