% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_strings.R
\name{new_definition}
\alias{new_definition}
\title{Definition Constructor}
\usage{
new_definition(x, dist)
}
\arguments{
\item{x}{List with the elements of the definition:
\code{dtDefs}, \code{varname}, \code{formula}, \code{param1}, \code{param2}}

\item{dist}{String with the distribution name.}
}
\value{
S3 object with class dist.
}
\description{
Definition Constructor
}
\concept{custom_distributions}
