% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_rmult.R
\name{rmult}
\alias{rmult}
\title{Internal function genExposure - returns categorical data}
\usage{
rmult(p)
}
\arguments{
\item{p}{a vector of probabilities}
}
\value{
An integer (group) ranging from 1 to length of the probability vector
}
\description{
Internal function genExposure - returns categorical data
}

