% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy-package.R
\name{simstudy-deprecated}
\alias{simstudy-deprecated}
\title{Deprecated functions in simstudy}
\description{
These functions are provided for compatibility with older versions
of simstudy only, and will be defunct in the future.
}
\details{
\itemize{
\item \link{genCorOrdCat}: This function is deprecated, and will
be removed in the future. Use \link{genOrdCat} with \code{asFactor = FALSE} instead.
\item \link{catProbs}: This function is deprecated, and will be removed in the future.
Use \link{genCatFormula} with the same functionality instead.
}
}
