% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simstudy-package.R
\docType{package}
\name{simstudy-package}
\alias{simstudy}
\alias{simstudy-package}
\title{simstudy: Simulation of Study Data}
\description{
Simulates data sets in order to explore modeling techniques or better understand data generating processes. The user specifies a set of relationships between covariates, and generates data based on these specifications. The final data sets can represent data from randomized control trials, repeated measure (longitudinal) designs, and cluster randomized trials. Missingness can be generated using various mechanisms (MCAR, MAR, NMAR).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kgoldfeld/simstudy}
  \item \url{https://kgoldfeld.github.io/simstudy/}
  \item \url{https://kgoldfeld.github.io/simstudy/dev/}
  \item Report bugs at \url{https://github.com/kgoldfeld/simstudy/issues}
}

}
\author{
\strong{Maintainer}: Keith Goldfeld \email{keith.goldfeld@nyulangone.org} (\href{https://orcid.org/0000-0002-0292-8780}{ORCID})

Authors:
\itemize{
  \item Keith Goldfeld \email{keith.goldfeld@nyulangone.org} (\href{https://orcid.org/0000-0002-0292-8780}{ORCID})
  \item Jacob Wujciak-Jens \email{jacob@wujciak.de} (\href{https://orcid.org/0000-0002-7281-3989}{ORCID})
}

}
\keyword{internal}
