% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{survGetParams}
\alias{survGetParams}
\title{Get survival curve parameters}
\usage{
survGetParams(points)
}
\arguments{
\item{points}{A list of two-element vectors specifying the desired time and 
probability pairs that define the desired survival curve}
}
\value{
A vector of parameters that define the survival curve optimized for
the target points. The first element of the vector represents the "f"
parameter and the second element represents the "shape" parameter.
}
\description{
Get survival curve parameters
}
\examples{
points <- list(c(60, 0.90), c(100, .75), c(200, .25), c(250, .10))
survGetParams(points)
}
\concept{utility}
