% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trtAssign.R
\name{trtAssign}
\alias{trtAssign}
\title{Assign treatment}
\usage{
trtAssign(dtName, nTrt = 2, balanced = TRUE, strata = NULL,
  grpName = "trtGrp")
}
\arguments{
\item{dtName}{data table}

\item{nTrt}{number of treatment groups}

\item{balanced}{indicator for treatment assignment process}

\item{strata}{vector of strings representing strativying variables}

\item{grpName}{string representing variable name for treatment or
exposure group}
}
\value{
An integer (group) ranging from 1 to length of the
probability vector
}
\description{
Assign treatment
}

