% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trtObserve.R
\name{trtObserve}
\alias{trtObserve}
\title{Observed exposure or treatment}
\usage{
trtObserve(dt, formulas, logit.link = FALSE, grpName = "trtGrp")
}
\arguments{
\item{dt}{data table}

\item{formulas}{collection of formulas that determine probabilities}

\item{logit.link}{indicator that specifies link. If TRUE, then logit link
is used. If FALSE, the identity link is used.}

\item{grpName}{character string representing name of treatment/exposure group
variable}
}
\value{
An integer (group) ranging from 1 to length of the probability vector
}
\description{
Observed exposure or treatment
}

