% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_strings.R
\name{validate}
\alias{validate}
\title{Validate a definition}
\usage{
validate(def)
}
\arguments{
\item{def}{S3 definition object with class `dist`.}
}
\description{
Each distribution has a validate method, that throws an error
 if the definition is not valid. Silent if valid.
}
\seealso{
[new_definition()]
}
\concept{custom_distributions}
