% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addColumns.R
\name{addColumns}
\alias{addColumns}
\title{Add columns to existing data set}
\usage{
addColumns(dtDefs, dtOld)
}
\arguments{
\item{dtDefs}{Name of definitions for added columns}

\item{dtOld}{Name of data table that is to be updated}
}
\value{
An updated data.table that contains the added simulated data
}
\description{
Add columns to existing data set
}

