% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defData.R
\name{defData}
\alias{defData}
\title{Add single row to definitions table}
\usage{
defData(dtDefs = NULL, varname = "", formula = "", variance = 0,
  dist = "Normal", link = "identity", id = "id")
}
\arguments{
\item{dtDefs}{Definition data.table to be modified}

\item{formula}{test}

\item{variance}{test}

\item{dist}{test}

\item{link}{test}

\item{id}{test}

\item{varName}{test}
}
\value{
A data.table named dtName that is an updated data defnitions table
}
\description{
Add single row to definitions table
}

