% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defMiss.R
\name{defMiss}
\alias{defMiss}
\title{Add single row to definitions table}
\usage{
defMiss(dtDefs = NULL, varname = "", formula = "", link = "identity",
  baseline = FALSE, monotonic = TRUE)
}
\arguments{
\item{dtDefs}{Definition data.table to be modified}

\item{formula}{test}

\item{link}{test}

\item{varName}{test}

\item{variance}{test}

\item{dist}{test}

\item{missType}{test}
}
\value{
A data.table named dtName that is an updated data defnitions table
}
\description{
Add single row to definitions table
}

