% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defData.R
\name{distributions}
\alias{distributions}
\alias{normal,}
\alias{poisson,}
\alias{noZeroPoisson,}
\alias{binary,}
\alias{binomial,}
\alias{uniform,}
\alias{categorical,}
\alias{gamma,}
\alias{beta,}
\alias{negBinomial,}
\alias{nonrandom,}
\alias{exponential,}
\alias{mixture,}
\alias{pseudorandom,}
\alias{pseudorandomSeq}
\title{Distributions for Data Definitions}
\arguments{
\item{formula}{Desired mean as a Number or an R expression for mean as a
String. Only variables defined via \code{\link{defData}} can be used within
the formula.}

\item{variance}{Number. Default is \code{0}.}

\item{link}{String identifying the link function to be used. Default is
\code{identity}.}
}
\description{
This help file describes the distributions used for data creation in
\code{simstudy}.
}
\details{
For details about the statistical distributions please see
 \code{\link[stats]{distribution}}, any non-statistical distributions will be
 explained below. Required variables and expected pattern for each
 distribution can be found in this table:

 \tabular{lllll}{ name \tab formula \tab format\tab variance \tab link \cr
 beta \tab mean \tab String or Number \tab dispersion value \tab  identity or
 logit \cr binary  \tab probability for 1 \tab String or Number \tab N.A.
 \tab  identity or logit \cr binomial\tab probability of success \tab String
 or Number \tab number of trials \tab identity or logit \cr categorical \tab
 probabilities \tab "p_1;p_2;..;p_n" \tab N.A. \tab N.A. \cr exponential \tab
 mean (lambda) \tab String or Number \tab N.A. \tab identity or log \cr gamma
 \tab mean \tab String or Number \tab dispersion value \tab identity or log
 \cr mixture \tab formula \tab "x_1 | p_1 + x_2 | p_2 ... x_n | p_n" \tab
 N.A. \tab N.A. \cr negBinomial \tab mean \tab String or Number \tab
 dispersion value \tab identity or log \cr nonrandom \tab formula \tab String
 or Number \tab N.A. \tab N.A. \cr normal \tab mean \tab String or Number
 \tab variance \tab N.A.\cr noZeroPoisson \tab mean \tab String or Number
 \tab N.A. \tab identity or log \cr poisson \tab mean \tab String or Number
 \tab N.A.  \tab identity or log \cr pseudorandom \tab vector and repetitions
 \tab "vector;reps" \tab N.A. \tab identity or fill \cr pseudorandomSeq \tab
 vector and repetitions \tab "vector;reps" \tab N.A. \tab identity or fill
 \cr uniform \tab range \tab "from;to" \tab N.A. \tab N.A. \cr uniformInt
 \tab range \tab "from;to" \tab N.A. \tab N.A. \cr }
}
\section{seudo-random and pseudo-random-sequence}{
 These distributions repeat
 the elements of a numeric vector a certain number of times in a random
 order. With \code{pseudorandom} every element in \code{vector} will be
 repeated \code{reps} times in a completly random order.
 \code{pseudorandomSeq} will also repeat the elements \code{reps} times but
 will randomize in sequences of \code{length(vector)}. E.g. \code{formula =
 "1,2,3;2"} could look like \code{3,1,1,2,3,2} for \code{pseudorandom} but
 something like \code{3,1,2,1,3,2} for \code{pseudorandomSeq}. Both
 distributions expect \code{reps * length(vector) = n} where \code{n} is the
 number of rows in the simulated data. Setting \code{link = "fill"} will
 supress this requirement and adjust the output to length \code{n}.
}

\section{mixture}{
 The mixture distribution makes it possible to mix to
 previously defined distributions/variables. Each variable that should be
 part of the new distribution \code{x_1,...,X_n} is assigned a probability
 \code{p_1,...,p_n}. For more information see
 \url{https://www.rdatagen.net/post/adding-mixture-distributions-to-simstudy/}
}

