% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_correlated_data.R
\name{genBlockMat}
\alias{genBlockMat}
\title{Create a block correlation matrix}
\usage{
genBlockMat(
  nInds,
  nPeriods,
  rho_w,
  rho_b = 0,
  rho_a = NULL,
  r = NULL,
  decay = NULL
)
}
\arguments{
\item{nInds}{The number of units (individuals) in each period.}

\item{nPeriods}{The number periods that data are observed.}

\item{rho_w}{The within-period correlation coefficient between -1 and 1.}

\item{rho_b}{The between-period correlation coefficient between -1 and 1.}

\item{rho_a}{The between-period within individual auto-correlation coefficient 
between -1 and 1.}

\item{r}{The decay parameter if correlation declines over time}

\item{decay}{The decay type can be "exp" or "prop". The decay structure}
}
\value{
A block correlation matrix of size (nInds \* nPeriods) x (nInds \* nPeriods)
}
\description{
The function genBlockMat() generates correlation matrices that 
can accommodate clustered observations over time where the within-cluster 
between-individual correlation in the same time period can be different from the 
within-cluster between-individual correlation across time periods.The matrix
generated here can be used in function addCorGen().
}
\examples{
genBlockMat(nInds = 4, nPeriods = 3, rho_w = .8)
genBlockMat(nInds = 4, nPeriods = 3, rho_w = .8, rho_b = 0.5)
genBlockMat(nInds = 4, nPeriods = 3, rho_w = .8, rho_b = 0.5, rho_a = 0.7)

genBlockMat(nInds = 4, nPeriods = 3, rho_w = .8, r = .9, decay = "exp")
genBlockMat(nInds = 4, nPeriods = 3, rho_w = .8, r = .9, decay = "prop")

}
\concept{correlated}
