% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_correlated_data.R
\name{genCorOrdCat}
\alias{genCorOrdCat}
\title{Generate correlated ordinal categorical data}
\usage{
genCorOrdCat(
  dtName,
  idname = "id",
  adjVar = NULL,
  baseprobs,
  prefix = "grp",
  rho,
  corstr,
  corMatrix = NULL
)
}
\description{
This function is deprecated, please use \link{genOrdCat} instead.
}
\keyword{internal}
