% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Calling function to simulate data}
\usage{
genData(n, dtDefs = NULL, id = "id")
}
\arguments{
\item{n}{the number of observations required in the data set.}

\item{dtDefs}{name of definitions data.table/data.frame. If no definitions are provided
a data set with ids only is generated.}

\item{id}{The string defining the id of the record}
}
\value{
A data.table that contains the simulated data.
}
\description{
Calling function to simulate data
}

