% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Calling function to simulate data}
\usage{
genData(dtDefs, n)
}
\arguments{
\item{dtDefs}{name of definitions data.table/data.frame}

\item{n}{the number (integer) of observations required in the data set}
}
\value{
A data.table that contains the simulated data
}
\description{
Calling function to simulate data
}

