% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_generate.R
\name{generate}
\alias{generate}
\title{Internal function called by gendt}
\usage{
generate(args, n, dfSim)
}
\arguments{
\item{args}{One row from data definitions data.table}

\item{n}{The number of observations required in the data set}

\item{dt}{Incomplete simulated data.table}
}
\value{
A data.frame with the updated simulated data
}
\description{
Internal function called by gendt
}
\keyword{internal}

