% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMiss.R
\name{genMiss}
\alias{genMiss}
\title{Generate missing data}
\usage{
genMiss(dtName, missDefs, idvars, nPeriods = 0, periodvar = "")
}
\arguments{
\item{dtName}{Name of complete data set}

\item{missDefs}{Definitions of missingness}

\item{idvars}{Index variables}

\item{nPeriods}{Number of periods (if longitudinal data)}

\item{periodvar}{Name of variable that contains period}
}
\value{
Missing data matrix indexed by idvars (and period if relevant)
}
\description{
Generate missing data
}

